REFACTOR: NAV-150 - Move utils to utils package
- Add __init__ files to sub packages.
- Add new asset package.
- Rename headers of pages.
