+++ public_transit_viewer/pages/isolines.py
@@ -9,15 +9,15 @@ from streamlit_folium import st_folium  # type: ignore
 from streamlit_searchbox import st_searchbox  # type: ignore
 
 from public_transit_viewer import LOGO_PATH
-from public_transit_viewer.client import get_isolines, get_stop_suggestions
 from public_transit_viewer.components.form_components import (
     query_config_expandable,
     time_form_row,
 )
-from public_transit_viewer.utils import get_color_map_hex_value
+from public_transit_viewer.utils.client import get_isolines, get_stop_suggestions
+from public_transit_viewer.utils.color import get_color_map_hex_value
 
 st.set_page_config(
-    page_title="Naviqore - Iso Lines",
+    page_title="Naviqore - Isolines",
     page_icon=str(LOGO_PATH),
 )
